api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.3.5
    # via aiohttp
aiohttp==3.10.1
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
fonttools==4.53.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
kiwisolver==1.4.5
    # via matplotlib
langchain==0.2.12
    # via langchain-community
langchain-community==0.2.11
    # via -r requirements.in
langchain-core==0.2.29
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
langchain-openai==0.1.20
    # via -r requirements.in
langchain-text-splitters==0.2.2
    # via langchain
langgraph==0.2.2
    # via -r requirements.in
langgraph-checkpoint==1.0.2
    # via langgraph
langsmith==0.1.98
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.21.3
    # via dataclasses-json
matplotlib==3.9.1.post1
    # via -r requirements.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   langchain
    #   langchain-community
    #   matplotlib
    #   pandas
openai==1.40.1
    # via langchain-openai
orjson==3.10.6
    # via langsmith
packaging==24.1
    # via
    #   langchain-core
    #   marshmallow
    #   matplotlib
    #   pytesseract
pandas==2.2.2
    # via -r requirements.in
pillow==10.4.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pytesseract
pydantic==2.8.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via pydantic
pyparsing==3.1.2
    # via matplotlib
pytesseract==0.3.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langsmith
    #   tiktoken
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.32
    # via
    #   langchain
    #   langchain-community
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.7.0
    # via langchain-openai
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
